+++ tests/bearer_auth.rs
@@ -152,7 +152,7 @@ async fn test_bearer_auth() -> Result<()> {
     // init segment will be retrieved twice, one of those times to print the PSSH if it is present.
     let outpath = env::temp_dir().join("bearer_auth.mp4");
     DashDownloader::new("http://localhost:6666/mpd")
-        .with_auth_bearer(String::from("magic"))
+        .with_auth_bearer("magic")
         .verbosity(2)
         .download_to(outpath.clone()).await
         .unwrap();